feat(models): save entity in database and applying firsts validations
So, i have created new action and saved object in database, i added references to models but it was
not make a sense, tomorow i create a migrate to remove this. I put a simple form gem in this project
because i use it by default.

BREAKING CHANGE: Adding nested form and save entity in database